criação da functions.dados_gerais
